# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TreeKB API
  version: "2.3.0"
  description: |
    TreeKB is an AI-driven, self-evolving knowledge graph system. It ingests documents,
    extracts knowledge points (KPs), organises them into a hierarchical topic tree, and
    provides intelligent question-answering with citations.

    **Key concepts**
    - **Dataset** -- an isolated knowledge base with its own SQLite database.
    - **Node** -- a topic node in the hierarchical tree.
    - **Chunk / KP** -- an atomic knowledge point stored under a node.
    - **Schema node** -- a curated node that guides ingestion mapping.
    - **Decision** -- a pending human-review item produced by the KP decision engine.

    Most endpoints require the `X-Dataset-ID` header to select the active dataset.
    Endpoints under `/datasets`, `/health`, `/metrics`, and `/settings` do not.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  parameters:
    DatasetIdHeader:
      name: X-Dataset-ID
      in: header
      required: true
      description: UUID of the target dataset. Resolved by middleware; falls back to default dataset when omitted.
      schema:
        type: string
        format: uuid
  schemas:
...
    SimpleError:
      type: object
      properties:
        error:
          type: string

    # ── Dataset ───────────────────────────────────────────────────────────────
...
    TestCase:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        query:
          type: string
        assertion_type:
          type: string
        assertion_value:
          type: string
        enabled:
          type: integer
          enum: [0, 1]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Prompt entry ──────────────────────────────────────────────────────────
...
paths:
...
  /settings/llm:
    get:
      operationId: getLlmSettings
      tags: [Settings]
      summary: Get LLM configuration
      description: Returns the current LLM provider, model, embedding model, and provider availability flags. Does not require X-Dataset-ID.
      responses:
        "200":
          description: Current LLM settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                    enum: [openai, gemini]
                  model:
                    type: string
                  embeddingModel:
                    type: string
                  openaiConfigured:
                    type: boolean
                  geminiConfigured:
                    type: boolean
                  vertexai:
                    type: boolean

    post:
      operationId: updateLlmSettings
      tags: [Settings]
      summary: Update LLM configuration
      description: Changes the active LLM provider and/or model at runtime. Does not require X-Dataset-ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [openai, gemini]
                model:
                  type: string
                embeddingModel:
                  type: string
      responses:
        "200":
          description: Updated LLM settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                  model:
                    type: string
                  embeddingModel:
                    type: string
                  openaiConfigured:
                    type: boolean
                  geminiConfigured:
                    type: boolean
                  vertexai:
                    type: boolean
        "400":
          description: Invalid provider value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

  # ── Query / Q&A ───────────────────────────────────────────────────────────
...
  /ingest/queue/stats:
    get:
      operationId: getIngestionQueueStats
      tags: [Ingestion]
      summary: Get ingestion queue statistics
      description: Returns counts of jobs by status in the ingestion queue.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Queue statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  queued:
                    type: integer
                  processing:
                    type: integer
                  completed:
                    type: integer
                  failed:
                    type: integer
                  rate_limited:
                    type: integer
                  cancelled:
                    type: integer

  # ── Documents ─────────────────────────────────────────────────────────────
...
  /embeddings/coverage:
    get:
      operationId: getEmbeddingCoverage
      tags: [Embeddings]
      summary: Get embedding coverage stats
      description: Returns the number and percentage of chunks and nodes that have embeddings.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Coverage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunks:
                    type: object
                    properties:
                      total:
                        type: integer
                      embedded:
                        type: integer
                      coverage:
                        type: number
                        format: float
                  nodes:
                    type: object
                    properties:
                      total:
                        type: integer
                      embedded:
                        type: integer
                      coverage:
                        type: number
                        format: float

  # ── Stats ─────────────────────────────────────────────────────────────────
  /stats:
    get:
      operationId: getSystemStats
      tags: [Stats]
      summary: Get system statistics
      description: Returns comprehensive statistics including tree, document, chunk, embedding, conflict, entity extraction, and queue stats for the active dataset.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: System statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: object
                  documents:
                    type: object
                  chunks:
                    type: object
                    properties:
                      active:
                        type: integer
                  embeddings:
                    type: object
                  conflicts:
                    type: object
                    properties:
                      total:
                        type: integer
                      unresolved:
                        type: integer
                      resolved:
                        type: integer
                  extraction:
                    type: object
                    properties:
                      entities:
                        type: integer
                      facts:
                        type: integer
                      documents_extracted:
                        type: integer
                  queue:
                    type: object
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /stats/tokens:
    get:
      operationId: getTokenStats
      tags: [Stats]
      summary: Get token usage statistics
      description: Returns LLM token usage statistics, optionally filtered by date, operation, or model.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: since
          in: query
          description: ISO date string to filter from
          schema:
            type: string
            format: date-time
        - name: operation
          in: query
          description: Filter by operation type
          schema:
            type: string
        - name: model
          in: query
          description: Filter by model name
          schema:
            type: string
      responses:
        "200":
          description: Token usage statistics
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

    delete:
      operationId: cleanupTokenHistory
      tags: [Stats]
      summary: Clean up old token history
      description: Deletes token usage records older than the specified number of days.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: daysToKeep
          in: query
          description: Number of days of history to retain
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Cleanup result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted_records:
                    type: integer
  /tree:
    delete:
      operationId: emptyTree
      tags: [Stats]
      summary: Empty the entire tree
      description: |
        **Dangerous operation.** Deletes all nodes, chunks, documents, and embeddings for the active dataset.
        Requires `?confirm=yes` query parameter. This action cannot be undone.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: confirm
          in: query
          required: true
          description: Must be "yes" to confirm
          schema:
            type: string
            enum: ["yes"]
      responses:
        "200":
          description: Tree emptied
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  nodes_deleted:
                    type: integer
                  chunks_deleted:
                    type: integer
                  documents_deleted:
                    type: integer
        "400":
          description: Missing confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string

  # ── Decisions ─────────────────────────────────────────────────────────────
...
  /tests:
    get:
      operationId: listTestCases
      tags: [Tests]
      summary: List all test cases
      description: Returns all test cases for the current dataset.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Test case list
          content:
            application/json:
              schema:
                type: object
                properties:
                  test_cases:
                    type: array
                    items:
                      $ref: "#/components/schemas/TestCase"

    post:
      operationId: createTestCase
      tags: [Tests]
      summary: Create a test case
      description: Creates a new test case with a query, assertion type, and expected value.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, query, assertion_type]
              properties:
                name:
                  type: string
                description:
                  type: string
                query:
                  type: string
                assertion_type:
                  type: string
                  description: e.g. "contains", "equals", "regex"
                assertion_value:
                  type: string
      responses:
        "201":
          description: Test case created
          content:
            application/json:
              schema:
                type: object
                properties:
                  test_case:
                    $ref: "#/components/schemas/TestCase"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /tests/{id}:
    put:
      operationId: updateTestCase
      tags: [Tests]
      summary: Update a test case
      description: Updates one or more fields of an existing test case.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                query:
                  type: string
                assertion_type:
                  type: string
                assertion_value:
                  type: string
                enabled:
                  type: integer
                  enum: [0, 1]
      responses:
        "200":
          description: Test case updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  test_case:
                    $ref: "#/components/schemas/TestCase"
        "404":
          description: Test case not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

    delete:
      operationId: deleteTestCase
      tags: [Tests]
      summary: Delete a test case
      description: Permanently deletes a test case.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test case deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          description: Test case not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

  # ── Schema ────────────────────────────────────────────────────────────────
...